Collection view of all a cats pictures now scales to 100 height and fills in the whole image.  Almost done working on hi-lighting of selected cell and have push left and push right transitions working correctly.

+++ Feline Finder/RescuePetsAPI2.swift
@@ -59,7 +59,7 @@ struct cat: Codable {
     var animalSex: String?
     var animalPrimaryBreed: String?
     var animalOrgID: String?
-    var animalLocationDistance: Int?
+    var animalLocationDistance: Double?
     var animalStatus: String?
     var animalBirthdate: String?
     var animalAvailableDate: String?
@@ -176,7 +176,7 @@ struct cat: Codable {
         }
 
         do {
-            animalLocationDistance = try values.decode(Int.self, forKey: .animalLocationDistance)
+            animalLocationDistance = Double(try values.decode(Int.self, forKey: .animalLocationDistance))
         } catch {
             animalLocationDistance = nil
         }
